Implement rank/select for RRR
