+++ include/mtl/succinct/bit_vector.hpp
@@ -1,6 +1,7 @@
 #pragma once
 #include "../bit_manip.hpp"
 #include "select.hpp"
+#include "bv.hpp"
 #include <cstddef>
 #include <cstdint>
 #include <vector>
@@ -12,6 +13,7 @@
 struct Bitmap {
   using value_type = bool;
   using W = uint64_t;
+  using rs_type = BV<Bitmap>;
   std::vector<W> arr;
   size_t sz;
   static constexpr size_t required_word_size(size_t n) {
@@ -273,138 +275,11 @@ struct Bitmap {
       x |= arr[b/64+1] << (64-r);
     return x & mask;
   }
-};
-
-template<typename T>
-struct RankSelectSupport {};
-
-template<>
-struct RankSelectSupport<Bitmap> {
-  static constexpr unsigned L = 512;
-  static constexpr unsigned S = 64;
-  static constexpr unsigned S_PER_L = L / S;
-  Bitmap* bm_;
-  std::vector<uint64_t> _r, _s, _zs;
-  RankSelectSupport() = default;
-  RankSelectSupport(const Bitmap& bm) {
-    build(bm);
+  const uint64_t& get_word(size_t wi) const {
+    return arr[wi];
   }
-  void build(const Bitmap& bm) {
-    bm_ = const_cast<Bitmap*>(&bm);
-    const auto num_l = (bm.size() + L-1) / L;
-    _r.assign((num_l + 1) * 2, 0);
-    _s.clear();
-    _s.push_back(0);
-    _zs.clear();
-    _zs.push_back(0);
-    uint64_t sum = 0;
-    for (size_t l = 0; l < num_l; ++l) {
-      auto psum = sum;
-      uint64_t sum_l = 0;
-      for (size_t s = 0; s < S_PER_L; ++s) {
-        if (l * S_PER_L + s < bm.arr.size())
-          sum_l += bm::popcnt(bm.arr[l * S_PER_L + s]);
-        if (s < S_PER_L-1)
-          _r[l * 2 + 1] |= sum_l << ((7-(s+1)) * 9);
-      }
-      sum += sum_l;
-      _r[(l + 1) * 2] = sum;
-      if (psum / L != sum / L) {
-        _s.push_back(l);
-      }
-      if ((L*l - psum) / L != (L*(l+1) - sum) / L) {
-        _zs.push_back(l);
-      }
-    }
-    _s.push_back(num_l);
-    _zs.push_back(num_l);
-  }
-
-  template<bool B>
-  inline size_t get_l(size_t l) const {
-    return B ? _r[l * 2] : l * L - _r[l * 2];
-  }
-  template<bool B>
-  inline size_t get_s(size_t l, size_t s) const {
-    auto b = (_r[l*2+1] >> ((7-s)*9)) % (1ull<<9);
-    return B ? b : s * S - b;
-  }
-  size_t rank1(size_t i) const { 
-    auto l = i / L;
-    auto s = i % L / S;
-    auto q = i / S;
-    auto r = i % S;
-    return get_l<1>(l) + get_s<1>(l, s) + bm::popcnt(bm_->arr[q] & ((1ull<<r)-1));
-  }
-  size_t rank0(size_t i) const { 
-    return i - rank1(i);
-  }
-  template<bool B>
-  size_t rank(size_t i) const {
-    if constexpr (B)
-      return rank1(i);
-    else
-      return rank0(i);
-  }
-  template<bool B>
-  size_t select(size_t ith) const {
-    auto n = ith+1; // to 1-indexed
-    if (n > rank<B>(bm_->size()))
-      return bm_->size();
-    // Find L block
-    auto& idx = B ? _s : _zs;
-    size_t l = idx[n / L];
-    {
-      auto r = idx[n / L + 1] + 1;
-      while (l+1 < r) {
-        auto c = l + (r-l)/2;
-        if (get_l<B>(c) < n)
-          l = c;
-        else
-          r = c;
-      }
-    }
-    // Find S block
-    size_t s = 0;
-    auto m = n - get_l<B>(l);
-    /* Following bit-manipulates code is same as ... */
-//    {
-//      auto d = 8;
-//      while (d > 1) {
-//        auto c = s + d/2;
-//        if (get_s<B>(l, c) < m)
-//          s = c;
-//        d /= 2;
-//      }
-//    }
-    {
-      uint64_t x = (uint64_t) (m-1) * 0x0040201008040201ull;
-      uint64_t a = _r[l*2+1];
-      if (!B)
-        a = 0x10100C08050301C0ull - a; // to 0s sum
-      uint64_t xda = x - a;
-      uint64_t sm = 0x4020100804020100ull;
-      uint64_t ok = (~x | a) & sm;
-      uint64_t ng = (~x & a) & sm;
-      uint64_t y = ((x ^ a ^ xda) & ok) | ng;
-      y = y * 0x0001010101010101ull >> (64-1-7);
-      auto id = bm::clz8(y)-1;
-      auto clo = bm::clz((~xda << 1 | 1) << (9*id));
-      auto ns = id + (clo ? (clo - 1) / 9 : 0);
-      s = ns;
-    }
-    // Calc select
-    auto o = m - get_s<B>(l, s);
-    uint64_t w = bm_->arr[l * S_PER_L + s];
-    return l * L + 
-           s * S + 
-           select64::select<B>(o-1, w);
-  }
-  size_t select1(size_t ith) const {
-    return select<1>(ith);
-  }
-  size_t select0(size_t ith) const {
-    return select<0>(ith);
+  size_t word_size() const {
+    return arr.size();
   }
 };
 
@@ -413,13 +288,15 @@ struct BitVector {
   static constexpr unsigned S = 64;
   static constexpr unsigned S_PER_L = L / S;
   Bitmap bm;
-  using rs_type = RankSelectSupport<Bitmap>;
+  using rs_type = typename Bitmap::rs_type;
   rs_type rs_support;
   // std::vector<uint64_t> _r, _s, _zs;
 
   BitVector(size_t size = 0, bool bit = false) : bm(size, bit) {}
   template<typename It>
-  BitVector(It begin, It end) : bm(begin, end) {}
+  BitVector(It begin, It end) : bm(begin, end) {
+    build();
+  }
   size_t size() const { return bm.size(); }
   bool empty() const { return bm.empty(); }
   void push_back(bool bit) { bm.push_back(bit); }
@@ -435,108 +312,15 @@ struct BitVector {
   Bitmap::const_iterator cend() const { return bm.cend(); }
 
   void build() {
-    rs_support = rs_type(bm);
-    // const auto num_l = (bm.size() + L-1) / L;
-    // _r.assign((num_l + 1) * 2, 0);
-    // _s.clear();
-    // _s.push_back(0);
-    // _zs.clear();
-    // _zs.push_back(0);
-    // uint64_t sum = 0;
-    // for (size_t l = 0; l < num_l; ++l) {
-    //   auto psum = sum;
-    //   uint64_t sum_l = 0;
-    //   for (size_t s = 0; s < S_PER_L; ++s) {
-    //     if (l * S_PER_L + s < bm.arr.size())
-    //       sum_l += bm::popcnt(bm.arr[l * S_PER_L + s]);
-    //     if (s < S_PER_L-1)
-    //       _r[l * 2 + 1] |= sum_l << ((7-(s+1)) * 9);
-    //   }
-    //   sum += sum_l;
-    //   _r[(l + 1) * 2] = sum;
-    //   if (psum / L != sum / L) {
-    //     _s.push_back(l);
-    //   }
-    //   if ((L*l - psum) / L != (L*(l+1) - sum) / L) {
-    //     _zs.push_back(l);
-    //   }
-    // }
-    // _s.push_back(num_l);
-    // _zs.push_back(num_l);
+    rs_support.build(bm);
   }
-
-  // template<bool B>
-  // inline size_t get_l(size_t l) const {
-  //   return B ? _r[l * 2] : l * L - _r[l * 2];
-  // }
-  // template<bool B>
-  // inline size_t get_s(size_t l, size_t s) const {
-  //   auto b = (_r[l*2+1] >> ((7-s)*9)) % (1ull<<9);
-  //   return B ? b : s * S - b;
-  // }
 
   inline size_t rank(size_t i) const {
     return rs_support.rank1(i);
-    // auto l = i / L;
-    // auto s = i % L / S;
-    // auto q = i / S;
-    // auto r = i % S;
-    // return get_l<1>(l) + get_s<1>(l, s) + bm::popcnt(bm.arr[q] & ((1ull<<r)-1));
   }
 
   template<bool B>
   size_t select(size_t n) const {
     return rs_support.select<B>(n);
-//     ++n; // to 1-indexed
-//     if (n > (B ? rank(size()) : size() - rank(size())))
-//       return size();
-//     // Find L block
-//     auto& idx = B ? _s : _zs;
-//     size_t l = idx[n / L];
-//     {
-//       auto r = idx[n / L + 1] + 1;
-//       while (l+1 < r) {
-//         auto c = l + (r-l)/2;
-//         if (get_l<B>(c) < n)
-//           l = c;
-//         else
-//           r = c;
-//       }
-//     }
-//     // Find S block
-//     size_t s = 0;
-//     auto m = n - get_l<B>(l);
-//     /* Following bit-manipulates code is same as ... */
-// //    {
-// //      auto d = 8;
-// //      while (d > 1) {
-// //        auto c = s + d/2;
-// //        if (get_s<B>(l, c) < m)
-// //          s = c;
-// //        d /= 2;
-// //      }
-// //    }
-//     {
-//       uint64_t x = (uint64_t) (m-1) * 0x0040201008040201ull;
-//       uint64_t a = _r[l*2+1];
-//       if (!B)
-//         a = 0x10100C08050301C0ull - a; // to 0s sum
-//       uint64_t xda = x - a;
-//       uint64_t sm = 0x4020100804020100ull;
-//       uint64_t ok = (~x | a) & sm;
-//       uint64_t ng = (~x & a) & sm;
-//       uint64_t y = ((x ^ a ^ xda) & ok) | ng;
-//       y = y * 0x0001010101010101ull >> (64-1-7);
-//       auto id = bm::clz8(y)-1;
-//       auto clo = bm::clz((~xda << 1 | 1) << (9*id));
-//       auto ns = id + (clo ? (clo - 1) / 9 : 0);
-//       s = ns;
-//     }
-//     // Calc select
-//     auto o = m - get_s<B>(l, s);
-//     uint64_t w = bm.arr[l * S_PER_L + s];
-//     return l * L + 
-//            s * S + 
-//            select64::select<B>(o-1, w);
   }
 };
